requirements.txt
streamlit>=1.20
mysql-connector-python>=8.0
python-dotenv>=1.0
bcrypt>=4.0